Move BlazeTargetLocationSearchView to a separate file
